attach

+++ mtx-admin/src/main/java/com/mtx/system/controller/biz/SystemAttachController.java
@@ -2,11 +2,15 @@ package com.mtx.system.controller.biz;
 
 import com.baomidou.mybatisplus.plugins.Page;
 import com.mtx.common.base.BaseController;
+import com.mtx.common.util.base.FileUtil;
 import com.mtx.common.util.page.PageFactory;
 import com.mtx.common.util.page.PageInfoBT;
 import com.mtx.common.util.wrapper.WrapMapper;
 import com.mtx.common.util.wrapper.Wrapper;
 import com.mtx.system.common.bean.DeleteFileTask;
+import com.mtx.system.common.exception.BusinessException;
+import com.mtx.system.common.exception.ErrorCodeEnum;
+import com.mtx.system.common.file.UploadComponent;
 import com.mtx.system.dao.dto.SystemAttachDto;
 import com.mtx.system.dao.model.SystemAttach;
 import com.mtx.system.dao.vo.SystemAttachVo;
@@ -22,6 +26,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.io.File;
 import java.util.List;
 
 @Controller
@@ -84,6 +89,13 @@ public class SystemAttachController extends BaseController {
     @ResponseBody
     public Object delete(@PathVariable("id") int id) {
         SystemAttachVo systemAttachVo = systemAttachService.selectByIdWithLeft(id);
+        String destFilePath = UploadComponent.getAbsolutePaths(systemAttachVo.getFilePath());
+        File file = new File(destFilePath);
+        //文件已在java中打开，从而不能删除
+        boolean fileInUsed = FileUtil.fileInUsed(file);
+        if(!fileInUsed){
+            throw new BusinessException(ErrorCodeEnum.SYS99990100);
+        }
         int count = systemAttachService.deleteByPrimaryKey(id);
         deleteFileTask.deleteFileByPaths(systemAttachVo.getFilePath());
         return WrapMapper.wrap(count);
@@ -95,6 +107,15 @@ public class SystemAttachController extends BaseController {
     @ResponseBody
     public Object delete(String ids) {
         String[] filePaths = systemAttachService.selectForFilePaths(ids);
+        for(String path:filePaths){
+            String destFilePath = UploadComponent.getAbsolutePaths(path);
+            File file = new File(destFilePath);
+            boolean fileInUsed = FileUtil.fileInUsed(file);
+            if(!fileInUsed){
+                throw new BusinessException(ErrorCodeEnum.SYS99990100);
+            }
+        }
+
         int count = systemAttachService.deleteByPrimaryKeys(ids);
         deleteFileTask.deleteFileByPaths(filePaths);
         return WrapMapper.wrap(count);
